Split integration tests by action
There is an integration test called test_person_resource.
This is a very long file and will keep growing.
Therefore split it into multiple files by action
