--- tests/integration_tests/__init__.py
@@ -1,1 +1,0 @@
-from tests.integration_tests.test_person_resource import *
